+++ changelog.md
@@ -1,13 +1,15 @@
 == Changelog ==
 
-= Version 3.2 – June 19, 2025 =
+= Version 3.2 – June 20, 2025 =
 
-+   New `eacKeyValue` helper class for key-value pair storage.
++   New `eacKeyValue` helper class for key/value pair storage.
     +   see: https://github.com/EarthAsylum/eacKeyValue
 +   Internal transient methods use key/value helper instead of WP transient API.
 +   Internal options methods use key/value helper instead of WP options API.
++   Reserved options now default to autoload=false.
++   Reworked internal option backup/restore.
 +   Added 'Key/Value storage' as session manager option.
-+   Session 'transient' option uses WP transient functions (not eacDoojigger).
++   Session 'transient' option uses WP transient functions (not internal).
 +   Make sure we have FS_CHMOD_FILE/FS_CHMOD_DIR set in autoload.php.
 +   Added `doTask()` method and `do_cron_task` action to event_scheduler extension.
 +   Delay scheduling events until `init` action, allows routing to Action Scheduler.